Moving backgrounds into correct folder and updating imports
