+++ frontend/src/assets/Background/SignUpBeam.png
