+++ frontend/src/assets/Background/SymbolBackground.png
